Restructuring server, Adding build commands, Cleaning up TS
