+++ package.json
@@ -21,6 +21,7 @@
   "scripts": {
     "aws-deploy": "npm run deploy -w aws",
     "aws-remove": "npm run remove -w aws",
+    "build-server": "npm run build -w server",
     "install-git-hooks": "npx simple-git-hooks",
     "lint": "npm run lint-client && npm run lint-server",
     "lint-client": "npm run lint:fix -w client",
@@ -33,6 +34,7 @@
   },
   "devDependencies": {
     "@babel/preset-env": "^7.19.3",
+    "@types/node": "^18.8.1",
     "concurrently": "^7.4.0",
     "eslint-import-resolver-typescript": "^3.5.1",
     "eslint-plugin-import": "^2.26.0",
