+++ server/.eslintignore
@@ -1,3 +1,3 @@
-build
+dist
 prod
 node_modules
