+++ server/package.json
@@ -1,9 +1,10 @@
 {
   "name": "server",
   "version": "1.0.0",
-  "main": "index.ts",
+  "main": "src/index.ts",
   "scripts": {
-    "start": "nodemon index.ts",
+    "build": "npx tsc",
+    "start": "nodemon src/index.ts",
     "lint": "eslint . --ext \".js,.ts\"",
     "lint:fix": "npm run lint -- --fix",
     "test": "echo \"Error: no test specified\" && exit 1"
@@ -19,10 +20,7 @@
   "devDependencies": {
     "@types/cors": "^2.8.12",
     "@types/express": "^4.17.14",
-    "@types/node": "^18.7.19",
-    "nodemon": "^2.0.20",
-    "ts-node": "^10.9.1",
-    "typescript": "^4.8.3"
+    "nodemon": "^2.0.20"
   },
   "engines": {
     "node": ">=16"
