+++ server/tsconfig.json
@@ -1,7 +1,8 @@
 {
   "extends": "../tsconfig.json",
   "compilerOptions": {
+    "noEmit": false,
     "module": "commonjs",
-    "outDir": "./dist"
+    "outDir": "dist",
   }
 }
